requirements.txt
alabaster==0.7.12
autopep8==1.6.0
Babel==2.10.3
biopython==1.79
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cycler==0.11.0
dgl-cu113==0.8.2.post1
dglgo==0.0.1
docutils==0.18.1
e3nn==0.5.0
filelock==3.7.1
fonttools==4.34.4
huggingface-hub==0.8.1
idna==3.3
imagesize==1.4.1
importlib-metadata==4.12.0
isort==5.10.1
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.3
MarkupSafe==2.1.1
matplotlib==3.5.2
mpmath==1.2.1
networkx==2.8.4
numpy==1.23.1
numpydoc==1.4.0
opt-einsum==3.3.0
opt-einsum-fx==0.1.4
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
psutil==5.9.1
pycodestyle==2.8.0
pydantic==1.9.1
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.7.9
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.1
scipy==1.8.1
sentencepiece==0.1.96
six==1.16.0
snowballstemmer==2.2.0
Sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sympy==1.10.1
threadpoolctl==3.1.0
tokenizers==0.12.1
toml==0.10.2
torch==1.11.0+cu113
torch-cluster==1.6.0
torch-geometric==2.0.4
torch-scatter==2.0.9
torch-sparse==0.6.14
torch-spline-conv==1.2.1
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.64.0
transformers==4.20.1
typer==0.6.1
typing_extensions==4.3.0
urllib3==1.26.10
zipp==3.8.1
